Start of Tutorial:
The base of this repository was generated using the processor archetype from: https://cwiki.apache.org/confluence/display/NIFI/Maven+Projects+for+Extensions

The source code for the processor is located at ./nifi-geecon-tutorial-processors/src/main/java/org/apache/nifi/tutorial/geecon/processors/MyProcessor.java. It is annotated with inline comments explaining the different basic methods.

The nifi-geecon-tutorial-nar folder contains the maven pom explaining how to create the NiFi archive file for geecon-tutorial-bundle. When run, it creates a 'nar' file which can be dropped into the 'lib' folder of a NiFi instance to add it to that deployment.

Execute ./gitNext.sh to get the next commit in the tutorial.