If you just cloned this repo, use this command to go to the beginning of the tutorial: git checkout ae1f11477230ee052d2a6ef5c7c5c12bc408d3c8
In this step you will try modifying the FlowFile yourself. For ideas, try to think of simple ways to modify the content or attributes. Check out MyProcessor for inline comments.

If you're feeling more adventurous try to think of common Java libraries that already exist which can be re-used in the onTrigger method to modify the content or extract information to be put into an attribute (like reading metadata from an image).